requirements.txt
fastapi
uvicorn[standard]
pydantic
groq
python-dotenv
httpx
jinja2
python-multipart
pymupdf
Pillow
pytesseract
python-docx
reportlab
